requirements.txt
pandas
numpy
pyspark
tweepy
tensorflow
matplotlib
PyYAML